# Speed up TP message lines split and iteration.